Add files via upload
Full description of First project:
THE BEGINNING OF THE WAY.
At some point in my life, I realized that everything around me lends itself to some kind of mathematical description. In all aspects of life, there are applications of computer technology that can make life a lot easier. I became very interested in the world of IT, which aims to use the computer as the main and powerful assistant in solving any problems. Since I am a biotechnologist by education, the profession closest to me related to IT is bioinformatics. 
I studied several computer languages (Python, SQL, R) to start my career in IT, and at some point it was time for my first project, where I applied the learned methods to determine the growth points of the company's product.