Analysis_of_4_companies
My second project involves analyzing a file of raw data. It involved looking at the relationships between the data, which I then had to analyze and draw conclusions from. The data is completely fictional and has no practical use, but this project is a good indicator of the skills I have learned so far in my IT training